Add files via upload
Remove duplicate code.
